test(server): isolate WebSocket contracts
Move the nine surviving WebSocket replay, live-stream, cursor, projection, and authentication contracts into one cohesive in-crate module. This keeps their private fixtures available and adds no integration link binary. Static leaf-name comparison against origin/main shows the one intentional deletion and no lost survivors.

Prune justification:
- ws_streams_a_turn_started_after_connecting: ws_replays_one_turn_then_streams_the_next_live already drives the same connect-before-live path while additionally pinning replay-to-live ordering and deduplication.
